src/django_backend/requirements.txt
django==2.1.1
djangorestframework
django-cors-headers==2.4.0
psycopg2-binary 
